change result of evaluating from double to int

+++ Calculator/evaluatingvisitor.cpp
@@ -19,7 +19,7 @@ void melihovv::calculator::Ast::EvaluatingVisitor::visit(
     const Addition* addition
     )
 {
-    std::tuple<double, double> t = visitBinaryOperation(addition);
+    std::tuple<int, int> t = visitBinaryOperation(addition);
     result = std::get<0>(t) + std::get<1>(t);
 }
 
@@ -27,7 +27,7 @@ void melihovv::calculator::Ast::EvaluatingVisitor::visit(
     const Subtraction* subtraction
     )
 {
-    std::tuple<double, double> t = visitBinaryOperation(subtraction);
+    std::tuple<int, int> t = visitBinaryOperation(subtraction);
     result = std::get<0>(t) - std::get<1>(t);
 }
 
@@ -35,7 +35,7 @@ void melihovv::calculator::Ast::EvaluatingVisitor::visit(
     const Multiplication* multiplication
     )
 {
-    std::tuple<double, double> t = visitBinaryOperation(multiplication);
+    std::tuple<int, int> t = visitBinaryOperation(multiplication);
     result = std::get<0>(t) * std::get<1>(t);
 }
 
@@ -43,7 +43,7 @@ void melihovv::calculator::Ast::EvaluatingVisitor::visit(
     const Division* division
     )
 {
-    std::tuple<double, double> t = visitBinaryOperation(division);
+    std::tuple<int, int> t = visitBinaryOperation(division);
     result = std::get<0>(t) / std::get<1>(t);
 }
 
@@ -51,7 +51,7 @@ void melihovv::calculator::Ast::EvaluatingVisitor::visit(
     const Power* power
     )
 {
-    std::tuple<double, double> t = visitBinaryOperation(power);
+    std::tuple<int, int> t = visitBinaryOperation(power);
     result = pow(std::get<0>(t), std::get<1>(t));
 }
 
@@ -60,14 +60,14 @@ double melihovv::calculator::Ast::EvaluatingVisitor::getResult() const
     return result;
 }
 
-std::tuple<double, double>
+std::tuple<int, int>
 melihovv::calculator::Ast::EvaluatingVisitor::visitBinaryOperation(
 const BinaryOperation* binaryOperation
 )
 {
     binaryOperation->leftNode()->accept(*this);
-    double left = result;
+    int left = result;
     binaryOperation->rightNode()->accept(*this);
-    double right = result;
+    int right = result;
     return std::make_tuple(left, right);
 }
