+++ Calculator/number.cpp
@@ -1,6 +1,6 @@
 #include "number.h"
 
-melihovv::calculator::Ast::Number::Number(double number)
+melihovv::calculator::Ast::Number::Number(int number)
 {
     this->number = number;
 }
@@ -10,7 +10,7 @@ void melihovv::calculator::Ast::Number::accept(Visitor& visitor) const
     visitor.visit(this);
 }
 
-double melihovv::calculator::Ast::Number::getNumber() const
+int melihovv::calculator::Ast::Number::getNumber() const
 {
     return number;
 }
